requirements.txt
torch
torchvision
timm
opencv-python
numpy
scikit-learn
pyyaml
Pillow
pandas